Added sandbox to the list of environments the log data queue is created in
This queue is used for s3 notifications for AWS config. As this wasn't
being created, some of the other resources which relied on it were
erroring.

I did think of removing the if condition altogether but there is a ddri
workspace on this account so I've added in the sandbox environment
instead.