Fix the existing modules.
The existing s3 modules were using tdr-terraform-modules which had tdr
specific parameter store names in it.

I tried to move it to the new da-terraform-modules but this didn't work
with the ancient aws provider this project is using.

I upgraded it and a bunch of things broke. It's now running a plan
without errors.